Added navigation `Link`s to `Header`
