+++ src/components/App.js
@@ -1,48 +1,20 @@
+import { Route, Switch } from 'react-router-dom';
 import Header from './Header';
-import Book from './Book';
-import AddBook from './AddBook';
+import Books from './Books';
+import Categories from './Categories';
 
-const App = () => {
-  const books = [
-    {
-      title: 'The Hunger Games',
-      author: 'Suzanne Collins',
-      genre: 'Action',
-      completed: 64,
-      id: 0,
-    },
-    {
-      title: 'Dune',
-      author: 'Frank Herbert',
-      genre: 'Science Fiction',
-      completed: 8,
-      id: 1,
-    },
-    {
-      title: 'Capital in the Twenty-First Century',
-      author: 'Suzanne Collins',
-      genre: 'Economy',
-      completed: 0,
-      id: 2,
-    },
-  ];
-  return (
-    <>
-      <Header />
-      <div>
-        {books.map((book) => (
-          <Book
-            key={book.id}
-            title={book.title}
-            author={book.author}
-            genre={book.genre}
-            completed={book.completed}
-          />
-        ))}
-      </div>
-      <AddBook />
-    </>
-  );
-};
+const App = () => (
+  <>
+    <Header />
+    <Switch>
+      <Route exact path="/">
+        <Books />
+      </Route>
+      <Route>
+        <Categories />
+      </Route>
+    </Switch>
+  </>
+);
 
 export default App;
