+++ src/components/Categories.js
@@ -1,0 +1,7 @@
+const Categories = () => (
+  <div>
+    No categories
+  </div>
+);
+
+export default Categories;
