+++ src/index.css
@@ -48,3 +48,8 @@ button {
   padding: 0.2em 0.8em;
   cursor: pointer;
 }
+
+a {
+  text-decoration: none;
+  color: inherit;
+}
